@@ -1,6 +1,6 @@
 class NewsBytesController < ApplicationController
   before_action :set_game, only: [:show, :edit, :update, :destroy]
-  before_filter :check_authorization, :except => 'index'
+  before_action :check_authorization, :except => 'index'
 
   # GET /news_bytes
   # GET /news_bytes.xml
