@@ -1,7 +1,7 @@
 class StandingsController < ApplicationController
   layout "application"
-  before_filter :check_authentication
-  before_filter :set_season
+  before_action :check_authentication
+  before_action :set_season
 
   def index
     @season_standings = Standing.season(:season_id => @season.id)
